Making max file size int

+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/internal/storage/Configuration.java
@@ -4,13 +4,13 @@ public class Configuration {
   public final long maxFileAgeForWriteInMillis;
   public final long minFileAgeForReadInMillis;
   public final long maxFileAgeForReadInMillis;
-  public final long maxFileSize;
+  public final int maxFileSize;
 
   public Configuration(
       long maxFileAgeForWriteInMillis,
       long minFileAgeForReadInMillis,
       long maxFileAgeForReadInMillis,
-      long maxFileSize) {
+      int maxFileSize) {
     this.maxFileAgeForWriteInMillis = maxFileAgeForWriteInMillis;
     this.minFileAgeForReadInMillis = minFileAgeForReadInMillis;
     this.maxFileAgeForReadInMillis = maxFileAgeForReadInMillis;
